Move super admin product maintanence page to superAdmin/products folder
Refs: PV-356
