+++ src/pages/superAdmin/products/CreateProduct.tsx
@@ -3,9 +3,9 @@ import { Notification } from 'hds-react';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
-import { makePrivate } from '../../auth/utils';
-import ProductForm from '../../components/superAdmin/products/ProductForm';
-import { MutationResponse } from '../../types';
+import { makePrivate } from '../../../auth/utils';
+import ProductForm from '../../../components/superAdmin/products/ProductForm';
+import { MutationResponse } from '../../../types';
 import styles from './CreateProduct.module.scss';
 
 const T_PATH = 'pages.superAdmin.createProduct';
