+++ src/pages/superAdmin/products/EditProduct.module.scss
