+++ src/pages/superAdmin/products/EditProduct.tsx
@@ -3,9 +3,9 @@ import { Notification } from 'hds-react';
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router';
-import { makePrivate } from '../../auth/utils';
-import ProductForm from '../../components/superAdmin/products/ProductForm';
-import { MutationResponse, Product } from '../../types';
+import { makePrivate } from '../../../auth/utils';
+import ProductForm from '../../../components/superAdmin/products/ProductForm';
+import { MutationResponse, Product } from '../../../types';
 import styles from './EditProduct.module.scss';
 
 const T_PATH = 'pages.superAdmin.editProduct';
